Fix : 수수료 관리 controller 병합 - 기존 dashboard 컨트롤러와 병합
