+++ ordering/src/main/java/com/beyond/MKX/domain/brokerage/controller/BrokerageDashboardController.java
@@ -2,17 +2,14 @@ package com.beyond.MKX.domain.brokerage.controller;
 
 import com.beyond.MKX.common.apiResponse.ApiResponse;
 import com.beyond.MKX.domain.account.admin.client.AdminInternalClient;
-import com.beyond.MKX.domain.brokerage.dto.BrokerageStatsDTO;
-import com.beyond.MKX.domain.brokerage.dto.LedgerDTO;
-import com.beyond.MKX.domain.brokerage.dto.OrderLogDTO;
-import com.beyond.MKX.domain.brokerage.dto.PopularStockDTO;
-import com.beyond.MKX.domain.brokerage.dto.RecentActivityDTO;
+import com.beyond.MKX.domain.brokerage.dto.*;
 import com.beyond.MKX.domain.brokerage.service.BrokerageDashboardService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
@@ -20,6 +17,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -29,7 +27,7 @@ import java.util.UUID;
  * - 증권사 관리자용 대시보드 데이터 조회
  */
 @RestController
-@RequestMapping("/api/brokerage/dashboard")
+@RequestMapping("/api/brokerage")
 @RequiredArgsConstructor
 @Slf4j
 public class BrokerageDashboardController {
@@ -43,7 +41,7 @@ public class BrokerageDashboardController {
      * @param adminId X-User-Id 헤더 (관리자 UUID)
      * @return 통계 데이터
      */
-    @GetMapping("/stats")
+    @GetMapping("/dashboard/stats")
     public ResponseEntity<?> getBrokerageStats(
             @RequestHeader(value = "X-User-Id", required = false) String adminId
     ) {
@@ -67,7 +65,7 @@ public class BrokerageDashboardController {
      * @param limit 최대 조회 개수 (기본값: 20)
      * @return 최근 활동 목록
      */
-    @GetMapping("/recent-activities")
+    @GetMapping("/dashboard/recent-activities")
     public ResponseEntity<?> getRecentActivities(
             @RequestHeader(value = "X-User-Id", required = false) String adminId,
             @RequestParam(defaultValue = "20") int limit
@@ -92,7 +90,7 @@ public class BrokerageDashboardController {
      * @param limit 최대 조회 개수 (기본값: 10)
      * @return 인기 종목 목록
      */
-    @GetMapping("/popular-stocks")
+    @GetMapping("/dashboard/popular-stocks")
     public ResponseEntity<?> getPopularStocks(
             @RequestHeader(value = "X-User-Id", required = false) String adminId,
             @RequestParam(defaultValue = "10") int limit
@@ -118,7 +116,7 @@ public class BrokerageDashboardController {
      * @param size 페이지 크기 (기본값: 20)
      * @return 주문 내역 페이지
      */
-    @GetMapping("/order-logs")
+    @GetMapping("/dashboard/order-logs")
     public ResponseEntity<?> getOrderLogs(
             @RequestHeader(value = "X-User-Id", required = false) String adminId,
             @RequestParam(defaultValue = "0") int page,
@@ -146,7 +144,7 @@ public class BrokerageDashboardController {
      * @param size 페이지 크기 (기본값: 20)
      * @return 거래내역 페이지
      */
-    @GetMapping("/ledgers")
+    @GetMapping("/dashboard/ledgers")
     public ResponseEntity<?> getLedgers(
             @RequestHeader(value = "X-User-Id", required = false) String adminId,
             @RequestParam(defaultValue = "0") int page,
@@ -163,6 +161,94 @@ public class BrokerageDashboardController {
         } catch (Exception e) {
             log.error("Failed to get ledgers for brokerage: {}", e.getMessage(), e);
             return ResponseEntity.badRequest().body(Map.of("message", "거래내역 조회 실패: " + e.getMessage()));
+        }
+    }
+
+
+
+    /**
+     * 일일 수수료 합산 조회 (당일 거래내역에서 합산)
+     * GET /api/brokerage/commission/daily
+     *
+     * @param adminId X-User-Id 헤더 (관리자 UUID)
+     * @param date 조회 날짜 (기본값: 오늘, 형식: yyyy-MM-dd)
+     * @return 일일 수수료 데이터
+     */
+    @GetMapping("/commission/daily")
+    public ResponseEntity<?> getDailyCommission(
+            @RequestHeader(value = "X-User-Id", required = false) String adminId,
+            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate date
+    ) {
+        try {
+            if (adminId == null || adminId.isBlank()) {
+                return ResponseEntity.badRequest().body(Map.of("message", "X-User-Id 헤더가 필요합니다."));
+            }
+            UUID brokerageId = getBrokerageId(UUID.fromString(adminId));
+            DailyCommissionDTO result = brokerageDashboardService.getDailyCommission(brokerageId, date);
+            return ApiResponse.ok(result, "일일 수수료 조회 성공");
+        } catch (Exception e) {
+            log.error("Failed to get daily commission: {}", e.getMessage(), e);
+            return ResponseEntity.badRequest().body(Map.of("message", "일일 수수료 조회 실패: " + e.getMessage()));
+        }
+    }
+
+    /**
+     * 기간별 수수료 합산 조회 (7일, 30일, 90일, 365일)
+     * GET /api/brokerage/commission/period
+     *
+     * @param adminId X-User-Id 헤더 (관리자 UUID)
+     * @param days 조회 기간 (7, 30, 90, 365) (기본값: 7)
+     * @return 기간별 수수료 데이터
+     */
+    @GetMapping("/commission/period")
+    public ResponseEntity<?> getCommissionByPeriod(
+            @RequestHeader(value = "X-User-Id", required = false) String adminId,
+            @RequestParam(required = false, defaultValue = "7") Integer days
+    ) {
+        try {
+            if (adminId == null || adminId.isBlank()) {
+                return ResponseEntity.badRequest().body(Map.of("message", "X-User-Id 헤더가 필요합니다."));
+            }
+
+            // days 유효성 검사
+            if (days != 7 && days != 30 && days != 90 && days != 365) {
+                return ResponseEntity.badRequest().body(
+                        Map.of("message", "days는 7, 30, 90, 365 중 하나여야 합니다.")
+                );
+            }
+
+            UUID brokerageId = getBrokerageId(UUID.fromString(adminId));
+            PeriodCommissionDTO result = brokerageDashboardService.getCommissionByPeriod(brokerageId, days);
+            return ApiResponse.ok(result, "기간별 수수료 조회 성공");
+        } catch (Exception e) {
+            log.error("Failed to get commission by period: {}", e.getMessage(), e);
+            return ResponseEntity.badRequest().body(Map.of("message", "기간별 수수료 조회 실패: " + e.getMessage()));
+        }
+    }
+
+    /**
+     * 시간대별 수익 추이 조회
+     * GET /api/brokerage/commission/hourly-trends
+     *
+     * @param adminId X-User-Id 헤더 (관리자 UUID)
+     * @param days 조회 기간 (기본값: 7일)
+     * @return 시간대별 수수료 추이 리스트
+     */
+    @GetMapping("/commission/hourly-trends")
+    public ResponseEntity<?> getHourlyCommissionTrends(
+            @RequestHeader(value = "X-User-Id", required = false) String adminId,
+            @RequestParam(required = false, defaultValue = "7") Integer days
+    ) {
+        try {
+            if (adminId == null || adminId.isBlank()) {
+                return ResponseEntity.badRequest().body(Map.of("message", "X-User-Id 헤더가 필요합니다."));
+            }
+            UUID brokerageId = getBrokerageId(UUID.fromString(adminId));
+            List<HourlyCommissionTrendDTO> trends = brokerageDashboardService.getHourlyCommissionTrends(brokerageId, days);
+            return ApiResponse.ok(trends, "시간대별 수익 추이 조회 성공");
+        } catch (Exception e) {
+            log.error("Failed to get hourly commission trends: {}", e.getMessage(), e);
+            return ResponseEntity.badRequest().body(Map.of("message", "시간대별 수익 추이 조회 실패: " + e.getMessage()));
         }
     }
